Update calculate_gains.py
Written basic code for calculating cost basis --> removed 'prev_cost_basis' and 'cur_cost_basis' as it was unnecessarily complex.  A for loop has been used but vectorisation would be preferred.  Next stage is to re-merge the buy_df with the sell_df and fill in the '0' cost basis column.